[hongxiang & mingheng] fix the bug that data precision loss for division operate

+++ src/main/java/com/linux/huhx/bootcamp/day_four_parkinglot/ParkingLot.java
@@ -43,6 +43,6 @@ public class ParkingLot {
   }
 
   public double getRemainSpaceRate() {
-    return getRemainSpace() / capacity;
+    return getRemainSpace() / (double) capacity;
   }
 }
